trafgen: l3: Support interface without IP address
Move default source address setting to packet_finish so that we do not
need to get the device's address if the source address is set in the
packet. Without this, trafgen does not with an interface without address
configured. In addition, in the case failing to get the address for an
interface, intead of panic, it now prints a warning and use a value of
0.0.0.0.

Signed-off-by: Ken Wu <abawwu@gmail.com>